evals/pyproject.toml
[project]
name = "evals"
version = "0.1.0"
description = ""
authors = [
    {name = "Devesh Meena",email = "nothefakedevesh@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai (>=2.16.0,<3.0.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "httpx (>=0.28.1,<0.29.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
